requirements.txt
Click==7.0
codecov==2.0.15
Flask==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pika==1.1.0
Pillow==7.1.1
protobuf==3.11.3
pytest==5.3.2
pytest-cov==2.8.1
requests==2.23.0